add support for yes-no, and strong/mild/no data
I removed Anna's workaround for a pandas error that comes up when the reader encounters a block of all missing categories. This happens for questions like 'Would you object if a family member brought a black friend to dinner?' (RACDIN), which they stopped asking in 1985.

The error has to be fixed in pandas.