draft of project proposal
I started putting this together using markdown and pandoc (check it out
if you're not familiar with it!). Feel free to change anything. At the
very least we need to include more information about the data sets (# of
features, missing data, etc), and also about our performance
measure (not sure what the best measure for time series performance is)