prejudice composite
currently living in 'raceData_withComposite.csv' as the prejudice variable. Should be easy to merge with existing data frames (via id and year). Sorry about doing this in R!

(btw this is an *interpretation* of the paper, not an exact replication. we can talk about this in our next meeting). It correlates with all the variables in expected ways though, so that's good.