moves some simple preprocessing to separate function
note that the zero-meaned year is in the column "year_norm", while the
true year is called "year".  The normed year is useful for the
regression because it's awkward for all the values to be so far from
zero when interpreting the resulting coefficients, but the true year is
needed for matching up court case dates.

Also, I've notived that the fit changes a lot every time I re-randomize
the training and test sets, so clearly we need to do stronger
regularization.