committing anna's ipynb changes, and possibly mine?
Ok so we need to stop committing to the same ipython notebooks, it
creates merge conflicts from hell.

Here's our idea: we each have a separate notebook that we work on, and
can also commit to and push if we want to show each other stuff, but
then when we've finished fiddling around with stuff, everything
permanent goes into regular python files, which should be easy to merge
when necessary.