testing all court windows from 1 to 10 years
Run each window 10 times with different train/test sets to get a good
estimate - takes a while. I saved an image of the results as
"summing_windows.png". The result is that short windows work a little
better for training error but all the window lengths seem to yield about
the same test error.